requirements.txt
accelerate==0.22.0
appdirs==1.4.4
autoencoder==0.0.8
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
contourpy==1.0.7
cycler==0.11.0
diffusers==0.10.2
docker-pycreds==0.4.0
einops==0.7.0
filelock==3.9.0
fonttools==4.39.4
fsspec==2023.9.0
ftfy==6.1.1
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.16.4
idna==3.4
imageio==2.30.0
importlib-metadata==6.8.0
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
lazy-loader==0.2
lit==15.0.7
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.2
pathtools==0.1.2
Pillow==9.3.0
protobuf==4.23.2
psutil==5.9.5
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.8.8
requests==2.28.1
safetensors==0.3.3
scikit-image==0.20.0
scipy==1.9.1
sentry-sdk==1.24.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
sympy==1.11.1
tifffile==2023.4.12
timm==0.9.16
tokenizers==0.13.3
torch==2.0.1
torch-fidelity==0.3.0
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
transformers==4.33.0
triton==2.0.0
typing-extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wandb==0.15.3
wcwidth==0.2.6
zipp==3.15.0
